-rustcrypto: Move implementations into modules
